upgrades

+++ functions.py
@@ -4,7 +4,7 @@ import pandas as pd
 
 from pathlib import Path
 
-def merge_files(path, delete=False):
+def merge_files(path, delete=False, precision='%0.4f'):
 
     path = Path(path)
 
@@ -16,14 +16,14 @@ def merge_files(path, delete=False):
         tmp = pd.read_csv(f, index_col=0)
         result = result.append(tmp)
 
-    result.sort_index().to_csv(fname, float_format='%0.3f')
+    result.sort_index().to_csv(fname, float_format=precision)
 
-    if (delete is True) & fname.exists():
+    if (delete is True) and fname.exists():
         for f in files:
             f.unlink()
 
 if __name__=='__main__':
 
-    path = '/work/GLEAM/perturbation_correction'
-    merge_files(path)
+    path = '/work/GLEAM/perturbation_correction_test_v5'
+    merge_files(path, precision = '%0.8f', delete=True)
